# caresplit_app/openapi.yaml
openapi: 3.0.3
info:
  title: CareSplit API
  version: 0.2.0
  description: >
    The backend contract the CareSplit frontend expects, derived directly
    from every call `frontend/` makes to `caresplit_backend.services.get_service()`
    (its `CareSplitService` protocol client — see
    `backend/caresplit_backend/services/base.py`). Implemented as a real,
    runnable FastAPI app at `backend/caresplit_backend/api/` (in-memory
    store, not a database) — see that package's `main.py` to run it. The
    frontend itself still talks to the service in-process, not over HTTP;
    this API is a standalone, independently-tested implementation of the
    same contract, not (yet) what the frontend actually calls.

    Scope note: only operations the frontend actually invokes are
    included. `CareSplitService` also defines `get_category(id)` and
    `get_expense(id)`, but no frontend page calls them — every page works
    from list results already in hand — so no `GET /categories/{id}` or
    `GET /expenses/{id}` is specified here. A real backend may still want
    to offer them, but the frontend does not require them.

    Authentication: every endpoint requires a bearer token **except**
    `POST /auth/token` (logging in, obviously, can't require the token
    logging in produces). Passwords are hashed server-side (PBKDF2-HMAC-
    SHA256 with a random salt, not reversible, never compared in plain
    text); tokens are opaque random strings held server-side with a
    24-hour expiry, not JWTs — there is nothing for a client to decode,
    and the server can't be tricked by a forged one. See
    `backend/caresplit_backend/api/auth.py`.

    Numbers: money and percentages are `Decimal` throughout the Python
    domain layer (see `backend/caresplit_backend/domain.py`), specifically
    to avoid float rounding error. They are serialized here as **strings**
    (`format: decimal`), not JSON numbers, so a client never has to round-
    trip them through a float. Field names are camelCase on the wire;
    the Python layer they come from is snake_case (e.g. `pctMe` ==
    `pct_me`).

    Receipts: `receiptFilename` is an opaque string the backend stores and
    returns verbatim. Today the frontend saves the uploaded file straight
    to local disk (`frontend/caresplit_frontend/receipts.py`) and only
    ever passes a filename through the service calls below — there is no
    upload endpoint in this contract, because the frontend never calls
    one.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
tags:
  - name: auth
  - name: categories
  - name: expenses
  - name: dashboard
  - name: reports
paths:
  /auth/token:
    post:
      tags: [auth]
      summary: Log in and obtain a bearer token
      description: >
        The only endpoint that does not require a bearer token — it's how
        you get one. Send it back as `Authorization: Bearer <accessToken>`
        on every other request.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Unknown username, or the password doesn't match it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /categories:
    get:
      tags: [categories]
      summary: List categories
      description: Called by the Expenses page (to build the category filter/select) and the Categories page.
      operationId: listCategories
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      tags: [categories]
      summary: Create a category
      description: Called from the Categories page's "Add category" dialog.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "409":
          description: >
            A category with this name already exists (case-insensitive).
            Maps to `DuplicateCategoryError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [categories]
      summary: Update a category
      description: >
        Called from the Categories page's edit dialog. The frontend always
        sends all three fields together today, but the underlying protocol
        (`update_category`) treats every field as optional, so this is
        specified as a partial update.
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: No category with this id. Maps to `NotFoundError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Renaming to a name that already exists. Maps to `DuplicateCategoryError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [categories]
      summary: Delete a category
      description: >
        Called from the Categories page's edit dialog. Fails if any
        expense still references this category — the frontend shows the
        409 message inline rather than pre-checking.
      operationId: deleteCategory
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "204":
          description: Deleted
        "404":
          description: No category with this id. Maps to `NotFoundError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Category still has expenses attached. Maps to `CategoryInUseError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /expenses:
    get:
      tags: [expenses]
      summary: List expenses, optionally filtered by category
      description: Called by the Expenses page for its main list and category filter.
      operationId: listExpenses
      parameters:
        - name: categoryId
          in: query
          required: false
          description: Omit for all expenses across every category.
          schema:
            type: integer
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: >
            OK. Newest first (by date, then id) — the frontend relies on
            this ordering and does not sort client-side.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      tags: [expenses]
      summary: Create an expense
      description: Called from the Expenses page's "Add expense" dialog.
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseCreate"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: categoryId does not refer to an existing category. Maps to `NotFoundError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /expenses/{expenseId}:
    parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [expenses]
      summary: Update an expense
      description: >
        Called from the Expenses page's edit dialog. The frontend always
        sends every field today (it re-submits the whole form), but the
        underlying protocol (`update_expense`) treats every field as
        optional, so this is specified as a partial update.
      operationId: updateExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: >
            No expense with this id, or (if categoryId was included in the
            request) categoryId does not refer to an existing category.
            Both map to `NotFoundError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [expenses]
      summary: Delete an expense
      description: Called from the Expenses page's edit dialog ("Delete" button).
      operationId: deleteExpense
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "204":
          description: Deleted
        "404":
          description: No expense with this id. Maps to `NotFoundError`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dashboard:
    get:
      tags: [dashboard]
      summary: This month's and this year's totals, plus recent expenses
      description: >
        Called once by the Home page. `today` is required: the frontend
        always computes its own local date and sends it explicitly
        (`date.today()`), rather than relying on the server's clock —
        deliberate, since a network-separated backend could be in a
        different timezone than the user.
      operationId: getDashboard
      parameters:
        - name: today
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTotals"
  /reports/years:
    get:
      tags: [reports]
      summary: Years that have at least one expense
      description: >
        Called once by the Reports page to populate the year picker.
        Descending order (newest first) — the frontend selects `years[0]`
        as the default and does not re-sort.
      operationId: listAvailableYears
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
  /reports/{year}:
    parameters:
      - name: year
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [reports]
      summary: Yearly report broken down by category and by month
      description: >
        Called by the Reports page whenever the selected year changes.
        Years with no expenses return a valid, empty report (zeroed
        totals, empty breakdowns) rather than a 404 — the frontend has no
        "no report for this year" error handling, only an empty-state
        message driven by empty arrays.
      operationId: getYearlyReport
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearlyReport"
components:
  responses:
    Unauthorized:
      description: Missing, invalid, or expired bearer token. Every operation except POST /auth/token can return this.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        An opaque token obtained from POST /auth/token, sent as
        `Authorization: Bearer <token>`. Not a JWT — the server holds the
        token/username/expiry mapping itself, so there's nothing for a
        client to inspect or decode.
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresAt]
      properties:
        accessToken:
          type: string
          description: "Opaque bearer token. Send as `Authorization: Bearer <accessToken>`."
        tokenType:
          type: string
          enum: [bearer]
        expiresAt:
          type: string
          format: date-time
          description: The token stops working after this instant (24 hours after issue).
    Decimal:
      type: string
      format: decimal
      description: A base-10 decimal, serialized as a string to avoid float rounding error. Matches Python's Decimal on the backend.
      example: "42.50"
    Category:
      type: object
      required: [id, name, pctMe, pctInsurance, notes]
      properties:
        id: { type: integer }
        name: { type: string }
        pctMe:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: Default percent of an expense in this category that is the user's share (0-100).
        pctInsurance:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: "Read-only, server-computed: 100 - pctMe."
          readOnly: true
        notes:
          type: string
    CategoryCreate:
      type: object
      required: [name, pctMe]
      properties:
        name: { type: string }
        pctMe: { $ref: "#/components/schemas/Decimal" }
        notes:
          type: string
          default: ""
    CategoryUpdate:
      type: object
      description: Every field optional; only the fields present are changed.
      properties:
        name: { type: string }
        pctMe: { $ref: "#/components/schemas/Decimal" }
        notes: { type: string }
    Expense:
      type: object
      required:
        - id
        - date
        - categoryId
        - categoryName
        - amount
        - pctMe
        - pctInsurance
        - amountMe
        - amountInsurance
        - description
        - note
        - receiptFilename
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        categoryId: { type: integer }
        categoryName:
          type: string
          description: Denormalized at write time from the category. Read-only.
          readOnly: true
        amount: { $ref: "#/components/schemas/Decimal" }
        pctMe:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: This expense's own split. Copied from the category's current pctMe at creation if not supplied; independent of the category afterward.
        pctInsurance:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: "Read-only, server-computed: 100 - pctMe."
          readOnly: true
        amountMe:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: "Read-only, server-computed: amount * pctMe / 100, rounded half-up to 2dp."
          readOnly: true
        amountInsurance:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: "Read-only, server-computed: amount - amountMe."
          readOnly: true
        description:
          type: string
          default: ""
        note:
          type: string
          default: ""
        receiptFilename:
          type: string
          nullable: true
          description: Opaque filename passed through as-is; the backend does not interpret it. No upload endpoint exists in this contract (see info.description).
    ExpenseCreate:
      type: object
      required: [date, categoryId, amount]
      properties:
        date: { type: string, format: date }
        categoryId: { type: integer }
        amount: { $ref: "#/components/schemas/Decimal" }
        description:
          type: string
          default: ""
        pctMe:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          nullable: true
          description: Omit or null to copy the category's current pctMe at creation time.
        note:
          type: string
          default: ""
        receiptFilename:
          type: string
          nullable: true
    ExpenseUpdate:
      type: object
      description: Every field optional; only the fields present are changed. pctMe is set literally (never re-derived from the category), including when categoryId also changes in the same request.
      properties:
        date: { type: string, format: date }
        categoryId: { type: integer }
        amount: { $ref: "#/components/schemas/Decimal" }
        description: { type: string }
        pctMe: { $ref: "#/components/schemas/Decimal" }
        note: { type: string }
        receiptFilename:
          type: string
          nullable: true
    Totals:
      type: object
      required: [total, me, insurance]
      properties:
        total: { $ref: "#/components/schemas/Decimal" }
        me: { $ref: "#/components/schemas/Decimal" }
        insurance:
          allOf: [{ $ref: "#/components/schemas/Decimal" }]
          description: "Read-only, server-computed: total - me."
          readOnly: true
    DashboardTotals:
      type: object
      required: [month, year, recent]
      properties:
        month:
          allOf: [{ $ref: "#/components/schemas/Totals" }]
          description: Totals for the calendar month containing the `today` query param.
        year:
          allOf: [{ $ref: "#/components/schemas/Totals" }]
          description: Totals for the calendar year containing the `today` query param.
        recent:
          type: array
          description: The 8 most recent expenses across all categories, newest first.
          maxItems: 8
          items:
            $ref: "#/components/schemas/Expense"
    CategoryBreakdown:
      type: object
      required: [categoryName, totals]
      properties:
        categoryName: { type: string }
        totals: { $ref: "#/components/schemas/Totals" }
    MonthBreakdown:
      type: object
      required: [month, totals]
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        totals: { $ref: "#/components/schemas/Totals" }
    YearlyReport:
      type: object
      required: [year, totals, byCategory, byMonth]
      properties:
        year: { type: integer }
        totals:
          allOf: [{ $ref: "#/components/schemas/Totals" }]
          description: Totals across the whole year.
        byCategory:
          type: array
          description: One entry per category that has at least one expense in this year, ordered by total descending. Categories with no expenses this year are omitted, not zeroed.
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        byMonth:
          type: array
          description: One entry per month (1-12) that has at least one expense, in calendar order. Months with no expenses are omitted, not zeroed.
          items:
            $ref: "#/components/schemas/MonthBreakdown"
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable message. The frontend displays this string directly in the dialog's error area — it is not machine-parsed.